fringematrix5-1yg: extract popovers to components and loadCampaignImages to useCampaignLoader hook
- Extract BuildInfoPopover and SharePopover into standalone components
  (client/src/components/BuildInfoPopover.tsx, SharePopover.tsx)
- Factor loadCampaignImages, selectCampaign, and all campaign loading
  state into useCampaignLoader hook (client/src/hooks/useCampaignLoader.ts)
- App.tsx wires up the hook and passes style/data props to popover components
- Update selectCampaign-cache-hit.test.js source guards to read from
  useCampaignLoader.ts (where the cache-hit logic now lives)
